[TASK] Add support for ajax search
# Support for huge amount of TYPO3 backend editors

This patch adds an AJAX based user search box. You
can search for username, realName, email and UID.

The ordering is based on lastlogin DESC, for now.

Furthermore, the user list becomes scrollable if it doesn't
fits into the viewport height of your browser.

# The default user list

is loaded if no search was performed or the search box
was cleared. The list is sorted by lastlogin DESC like the
search results.

# A custom CSS was introduced for visual improvements

- max-height of the user list is based von CSS "vh"
- user list items are built with flexbox
- a right-margin was introduced to not affect the scrollbar
  (feedback for improvements welcome)

Many thanks to Alex Nostadt for the inital PR!!!

Solves #5

Co-Authored-By: Alex Nostadt <amartinno1@users.noreply.github.com>