[WIP][TASK] Use new dependency injection
use the DI autowire feature whereever possible.

ToDo:
Checking whether an admin is in a switchuser backend session
must be changed because "be_sessions.ses_backuserid" can't
be accessed directly. According to
https://docs.typo3.org/c/typo3/cms-core/main/en-us/Changelog/11.0/Breaking-93023-ReworkedSessionHandling.html?highlight=reworked%20session%20handling
this must be gathered via UserSession's class get() or getAll().

Once the UserSession can be instantiated, the check if the beuser_fastswitch
toolbar item must be shown or not can be re-added.